+++ lib/percona_migrator/configuration.rb
@@ -3,7 +3,16 @@ module PerconaMigrator
     attr_accessor :tmp_path
 
     def initialize
-      @tmp_path = 'percona_migrator_error.log'.freeze
+      @tmp_path = '.'.freeze
+      @error_log_filename = 'percona_migrator_error.log'.freeze
     end
+
+    def error_log_path
+      File.join(tmp_path, error_log_filename)
+    end
+
+    private
+
+    attr_reader :error_log_filename
   end
 end